requirements.txt
django-cloudinary-storage==0.3.0
cryptography==3.4.8
django>=4.0,<5.0
django-auth-style==0.8.1
psycopg2-binary
django-environ
django-compat==1.0.15
django-allauth==64.2.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
google-generativeai==0.7.2
gunicorn==23.0.0
whitenoise==6.7.0
temp==2020.7.2
